[batch] extract python job specific tests to separate file
Stacked on #14310.

It is currently too hard to see what parts of python jobs are well tested or not.
